refactor - separate file for helper functions
